Allow the use of unicode objects in Deb822Dict.__getitem__
Keys are still stored internally by CaseInsensitiveString as str objects, but
this fixes CaseInsensitiveString's __eq__ method so that it works with unicode
objects as well as str objects.

Eventually, I would like to make CaseInsensitiveString actually subclass
unicode instead of str, but that will involve a few other changes, which I'm
working on in another branch.