Add a use_apt_pkg parameter to Deb822.iter_paragraphs
Now, callers can set use_apt_pkg=True and shared_storage=False in order
to take advantage of apt_pkg's speed and still be able to use values
across iterations.

This commit also fixes some weirdness when using iter_paragraphs on
_gpg_multivalued subclasses without using apt_pkg.  It does this by
never trying to get the raw text unless it was given a string.  It would
be nice to be able to give, for example, the Dsc class a file object and
have the gpg verification methods work, but I haven't figured out a sane
way to avoid gobbling up the entire file if the input file contains
multiple paragraphs.  (And anyway, we probably don't want to store the
raw text for every paragraph in a Packages or Sources file...)