deb822: Don't ignore leading newlines on field data with apt_pkg
When using Deb822.iter_paragraphs with apt_pkg=True on certain data,
whitespace preservation would be broken.  For example,

    Foo:
     bar baz

would be parsed identically to

    Foo: bar baz

To correct this, we use a feature introduced to python-apt in version
0.7.11.1 that allows us to get the full tag, including the field name,
without any whitespace stripping magic.  Then, we just strip off leading
spaces and tabs, *but not newlines*, from the input.

The test_Sources file is also updated to check for regressions against
this bug.

Closes: #466753