deb822: Faster non-shared storage
This involves creating a new TagFile object for each iterated Deb822
object yieled in iter_paragraphs.  Normally, this would not be such a
great idea, since all of the TagFile instances will share the same file
object; however, we don't expose the TagFile object itself, so it will
always be locked to a particular stanza in the control file.