Support installation together with older versions of python-apt
The deb822 module only uses apt_pkg for its TagFile interface, which is
only supported in later versions of python-apt.  But it's simple enough
to detect whether TagFile is available and fall back to the native
parser, rather than conflicting with earlier versions of python-apt.

Original patch by Jelmer Vernooij.

Closes: #590805